requirements.txt
torch==2.4.0
torchvision==0.19.0
transformers==4.44.2
Pillow==10.4.0
numpy==2.1.0
requests==2.32.3
scikit-image==0.24.0
